prints the largest prime factor of the number

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	long long int num, pfac;
+	long int num, pfac;
 
 	num = 612852475143;
 	for (pfac = 2; pfac <= num; pfac++)
@@ -19,6 +19,6 @@ int main(void)
 			pfac--;
 		}
 	}
-	printf("%lld\n", pfac);
+	printf("%ld\n", pfac);
 	return (0);
 }
